SONARIAC-846 Convert JSON to Minimal ARM AST Model: Refactor ArmParser
