+++ iac-common/src/test/resources/org/sonar/iac/common/checks/TestCheck.java
@@ -1,4 +1,4 @@
-package org.sonar.iac.common.checks
+package org.sonar.iac.common.checks;
 
 class TestCheck {
 }
