+++ iac-extensions/arm/src/main/java/org/sonar/iac/arm/tree/api/ParameterDeclaration.java
@@ -27,7 +27,7 @@ public interface ParameterDeclaration extends Statement {
 
   ParameterType type();
 
-  Expression defaultValue();
+  PropertyValue defaultValue();
 
   List<Expression> allowedValues();
