+++ iac-extensions/arm/src/main/java/org/sonar/iac/arm/tree/impl/json/ArmHelper.java
@@ -24,12 +24,20 @@ import java.util.Optional;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
 import org.sonar.iac.arm.tree.api.Expression;
+import org.sonar.iac.arm.tree.api.Property;
 import org.sonar.iac.arm.tree.api.SimpleProperty;
 import org.sonar.iac.common.api.tree.Tree;
 
 public class ArmHelper {
 
   private ArmHelper() {
+  }
+
+  public static void addChildrenIfPresent(List<Tree> children, @Nullable Property property) {
+    if (property != null) {
+      children.add(property.key());
+      children.add(property.value());
+    }
   }
 
   public static void addChildrenIfPresent(List<Tree> children, @Nullable SimpleProperty property) {
